+++ src/index.ts
@@ -1,2 +1,2 @@
-export { deepEqual } from './equal';
+export { deepEqual, deepEqualBy } from './equal';
 export type { DeepEqualOptions } from './types';